Concluído a criação de classes e métodos
