+++ exercicios_projeto/Client.js
@@ -23,39 +23,4 @@ class Clients {
     }
 }
 
-class Standard extends Clients {
-    constructor(name, cpf, email, telefone, monthlyIncome) {
-        if (monthlyIncome <= 4999.99) {
-            super(name, cpf, email, telefone, monthlyIncome);
-            this.transferLimit = 1000;
-            this.type = 'standard';
-        } else {
-            throw new Error('Não foi possível criar a conta Standard');
-        }
-    }
-}
-
-class Gold extends Clients {
-    constructor(name, cpf, email, telefone, monthlyIncome) {
-        if (monthlyIncome >= 5000 && monthlyIncome <= 17999.99) {
-            super(name, cpf, email, telefone, monthlyIncome);
-            this.transferLimit = 5000;
-            this.type = 'gold';
-        } else {
-            throw new Error('Não foi possível criar a conta Gold');
-        }
-    }
-}
-
-class Premium extends Clients {
-    constructor(name, cpf, email, telefone, monthlyIncome) {
-        if (monthlyIncome >= 18000) {
-            super(name, cpf, email, telefone, monthlyIncome);
-            this.type = 'premium';
-        } else {
-            throw new Error('Não foi possível criar a conta Premium');
-        }
-    }
-}
-
-module.exports = { Clients, Gold, Premium, Standard };
+module.exports = { Clients};
