+++ exercicios_projeto/PremiunAccount.js
@@ -1,0 +1,14 @@
+const { Clients } = require('./Client');
+
+class Premium extends Clients {
+    constructor(name, cpf, email, telefone, monthlyIncome) {
+        if (monthlyIncome >= 18000) {
+            super(name, cpf, email, telefone, monthlyIncome);
+            this.type = 'premium';
+        } else {
+            throw new Error('Não foi possível criar a conta Premium');
+        }
+    }
+}
+
+module.exports = { Premium };
