+++ exercicios_projeto/testes.test.js
@@ -1,0 +1,3 @@
+test('Exemplo de teste', () => {
+     expect(true).toBe(true);
+});
